解决目标选取父目录可以选择子孙栏目的问题与有子栏目删除问题的修复

+++ dream/admin/catedel.php
@@ -14,11 +14,33 @@
 删除cat_id
 */
 
+
+
 define('ACC', true);
 require('../include/init.php');
 $cat_id = $_GET['cat_id'] + 0; //“+0代码健壮性”
 
+/*
+判断该栏目是否有子栏目
+如果有子栏目，则该栏目不允许删除
+
+思路：
+无限级分类有3个基本应用
+1：查子栏目
+2：查子孙栏目
+3：查家谱树
+
+我们可以在model写一个方法，专门查子栏目
+*/
+
+
+
 $cat = new CatModel();
+
+$sons = $cat->getSon($cat_id); //判断有没有子栏目，有子栏目不允许删除
+if(!empty($sons)){
+	exit('有子栏目，不允许删除');
+}
 
 if($cat->delete($cat_id)){
 	echo '删除成功';
